requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tensorflow==2.9.1
torch==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0

# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
pytest>=7.0
python-dotenv>=0.5.1
